# libs/contracts/rest/openapi.snapshot.json
{
...
  "info": {
    "contact": {},
    "description": "Backend/API-first social monitoring MVP.",
    "title": "Social Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/scan-requests/{scanJobId}/status": {
      "get": {
        "operationId": "ScanStatusController_get",
        "parameters": [
          {
            "in": "path",
            "name": "scanJobId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:topics. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Scan job reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get current scan job status.",
        "tags": [
          "scan-requests"
        ]
      }
    },
...
    "/source-bindings/{sourceBindingId}/scan-requests": {
      "post": {
        "operationId": "ScanRequestController_create",
        "parameters": [
          {
            "in": "path",
            "name": "sourceBindingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:scan_requests. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Manual scan requests require owner, admin or member. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Request a scan for a source binding.",
        "tags": [
          "scan-requests"
        ]
      }
    },
...
  },
  "servers": [],
...
}